Move Sharing feature to vertical-slice folders
Pilot continues with the Sharing slice (5 endpoints — create,
revoke, list-mine, get-info-by-link, stream-by-link). Controller,
service, interface, DTOs, route constants, and tests now live
under Features/Sharing/ in their respective projects.

- ISharedLinkService moved out of Domain.Services into
  Application.Features.Sharing.
- ShareRoutes extracted from ApiEndpoints.cs.
- DI registration moved into per-feature AddSharingFeature()
  extension in SharingModule.cs.
- Mobile (Library, MAUI app, test project) gain a using for the
  new Contracts.Features.Sharing namespace alongside their
  existing Contracts.Requests/Responses usings.
