+++ src/backend/Application/Features/Sharing/SharingModule.cs
@@ -1,0 +1,15 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Application.Features.Sharing;
+
+public static class SharingModule
+{
+    extension(IServiceCollection services)
+    {
+        public IServiceCollection AddSharingFeature()
+        {
+            services.AddScoped<ISharedLinkService, SharedLinkService>();
+            return services;
+        }
+    }
+}
