+++ src/backend/TB.DanceDance.API.Contracts/Features/Sharing/CreateSharedLinkRequest.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace TB.DanceDance.API.Contracts.Requests
+namespace TB.DanceDance.API.Contracts.Features.Sharing
 {
     public class CreateSharedLinkRequest
     {
